requirements.txt
gnureadline==6.3.3
ipython==2.3.1
isodate==0.5.1
pywinrm==0.0.3
wsgiref==0.1.2
xmltodict==0.9.1
